requirements.txt
anticaptchaofficial==1.0.48
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
greenlet==2.0.1
idna==3.4
importlib-metadata==6.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
playwright==1.32.1
py==1.11.0
pyee==9.0.4
Pygments==2.15.0
python-anticaptcha==1.0.0
requests==2.28.2
rich==13.3.3
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.15
websocket-client==1.3.3
websockets==11.0.1
Werkzeug==2.2.3
zipp==3.15.0
yaspin~=2.3.0
